create two-page layout
the way this is set up, we can follow Jason's idea of having the 'info'
page on the root index page, and be able to set and control things from
the /control panel page. For Spotify, Kent and I may also add in some
extra pages accessable from the control panel page.